requirements.txt
# Web Framework & Server
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests

# AI Orchestration & Observability
langgraph
langsmith
openai

# Media & Utilities
Pillow
PyYAML

# RAG (document upload + story memory)
faiss-cpu
numpy
PyMuPDF
python-multipart

# PDF export
reportlab

# Testing
pytest
pytest-asyncio